agents/trump-melania/pyproject.toml
[tool.poetry]
name = "trump-melania-agent"
version = "1.0.0"
description = "Melania Trump agent"
authors = ["Coral Protocol"]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.20"
langchain-mcp-adapters = "^0.1.3"
langchain-openai = "^0.3.0"
python-dotenv = "^1.0.0"
solana = "^0.35.0"
solders = "^0.21.0"
base58 = "^2.1.1"
certifi = "^2024.0.0"
aiohttp = "^3.9.0"
requests = "^2.31.0"
cdp-sdk = "^0.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

